requirements.txt
langchain==0.0.216
langchainplus-sdk==0.0.17
chromadb==0.3.26
python-dotenv
discord.py==2.3.1
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-api-python-client==2.86.0
wikipedia==1.4.0
openai==0.27.4